4.4.5: The Doctor is In
Heyo, I am here to announce the introduction of Litchi Faye Ling to the Roster!

I have done the grouping of commands like I did with Makoto, the commands are m(Staff), and e(Staffless) like it was listed under the wiki.  I have also include reach as a sub group under the e commands including astral.

P.S. I will be making a wiki entry on how to fully use the commands and special cases if they come.  Sorry for the wait!