Bnscoffee
It has come to my attention recently that BNSCoffee has been shutdown, meaning the character search does not work,  I have plans at some point to revisit this and update it to use the BNS in house character search.  I have moved the Bnschar to another Branch for a future sitdown to update it.