Post Break fixes and update to HoTS Cog
So I forgot to put in that it requires tabulate in the json file for BBCF and HoTS, so I have added that to it.  I have also added new characters to the HoTS Cog and did the rework to the info as stuff was taken away and it wasn't working :( all good now though.