BBCF 1.5
It is now time, I have added in Jubei as of 9/24/17, since this character is new, I suspect changes with be happening so I will be checking the page often to make sure that Jubei is working.  Sorry for the delay, took a sanity break preparing for TFC next week!